+++ src/Components/Cryptocurrency.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Templateless\Components;
+
+enum Cryptocurrency: string
+{
+    case BITCOIN = 'bitcoin';
+    case ETHEREUM = 'ethereum';
+}