+++ src/Components/SignatureFont.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Templateless\Components;
+
+enum SignatureFont: string
+{
+    case REENIE_BEANIE = 'REENIE_BEANIE';
+    case MEOW_SCRIPT = 'MEOW_SCRIPT';
+    case CAVEAT = 'CAVEAT';
+    case ZEYADA = 'ZEYADA';
+    case PETEMOSS = 'PETEMOSS';
+}