+++ src/Components/StoreBadge.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Templateless\Components;
+
+enum StoreBadge: string
+{
+    case APP_STORE = 'APP_STORE';
+    case GOOGLE_PLAY = 'GOOGLE_PLAY';
+    case MICROSOFT_STORE = 'MICROSOFT_STORE';
+}