+++ CHANGELOG.md
@@ -1,15 +1,23 @@
 # CHANGELOG
 
-**v0.1.0-alpha.3:**
+## v0.1.0-alpha.4:
+
+- New social icons: `Service::MASTODON` and `Service::RSS`
+- New store badges component
+- New QR code component
+- New signature component
+- New [examples](examples)
+
+## v0.1.0-alpha.3:
 - `README.md`: notice about test mode
 - Support for test mode logging
 
-**v0.1.0-alpha.2:**
+## v0.1.0-alpha.2:
 - `README.md` cleanup (listing of components)
 - `Image` component now requires only `src`; the other params are optional
 - `ViewInBrowser` component has changed: text is optional
 
-**v0.1.0-alpha.1:**
+## v0.1.0-alpha.1:
 - Introduced `CHANGELOG.md`
 - Introduced new services as social icons:
   - `Service::PHONE` (converts into a link with `tel:` prefix)
@@ -24,5 +32,5 @@
   - `Service::THREADS`
   - `Service::TELEGRAM`
 
-**v0.1.0-alpha.0:**
+## v0.1.0-alpha.0:
 - Initial implementation